Swap VCC and GNC pins on BME680
This mistake nearly cost me the sensor. To my luck only the regulator
released the magick smoke. After replacing the regulator the sensor
started working again.

Also, the screen diode draws a lot of current. Through experimentation I
figured out that it's bright enough with just 15mA of current. So I
added a 440 Ohm load in series with the backlight to limit the current
to about that value. This made also all sensor readings more stable, as
the MH-Z19 sensor draws up-to 100mA which brown-outs the MCU and other
sensors. Lowering the LCD current draw fixed that problem as well.